Moved test data. Edited description.
