+++ tests/repositories/test-data.js
@@ -1,0 +1,37 @@
+const allUsersExample = [
+  {
+    id: 'd40ddf50-386e-4d6c-a10a-2c08d599ab19',
+    name: 'Software Engineering School',
+    email: 'software@engineering.school',
+    password: '$2a$08$qT85kN0AQU1mnl3fCyskzOfw/qxuQBRmzzIP1qwlB0QohZ0OghqSu',
+  },
+];
+
+const userExample = {
+  id: 'd40ddf50-386e-4d6c-a10a-2c08d599ab19',
+  name: 'Software Engineering School',
+  email: 'software@engineering.school',
+  password: '$2a$08$qT85kN0AQU1mnl3fCyskzOfw/qxuQBRmzzIP1qwlB0QohZ0OghqSu',
+};
+
+const candidateExample = {
+  name: 'Software Engineering School',
+  email: 'software@engineering.school',
+  password: 'ses12345',
+};
+
+const newUserExample = {
+  id: 'd40ddf50-386e-4d6c-a10a-2c08d599ab19',
+  name: 'Software Engineering School',
+  email: 'software@engineering.school',
+};
+
+const idExample = 'd40ddf50-386e-4d6c-a10a-2c08d599ab19';
+
+module.exports = {
+  allUsersExample,
+  userExample,
+  candidateExample,
+  newUserExample,
+  idExample,
+};
